requirements.txt
websockets
msgpack
Pillow
